+++ lib/canopus/workspace/language_server_configurable.rb
@@ -94,6 +94,7 @@ module Canopus
       if client
         @clients.delete(language)
         invalidate_diagnostics
+        invalidate_document_highlights(client: client)
         invalidate_inlay_hints(client: client)
         invalidate_code_lenses(client: client)
         invalidate_sticky_symbols(client: client)